refactor: create helpers for getGifs
