+++ src/helpers/getGifs.js
@@ -1,0 +1,16 @@
+export const getGifs = async(category) => {
+  const url = `https://api.giphy.com/v1/gifs/search?api_key=SehYrvVMKCk9VRuh9nyEK2w66JLEITpL&q=${encodeURI(category)}
+                &limit=10`;
+  const resp = await fetch(url);
+  const { data } = await resp.json();
+
+  const gifs = data.map(img => {
+    return {
+      id: img.id,
+      title: img.title,
+      url: img.images?.downsized_medium.url
+    };
+  });
+
+  return gifs;
+};
